Перевел суммы и int на float64.

+++ internal/accrualclient/reqres.go
@@ -1,7 +1,7 @@
 package accrualclient
 
 type Response struct {
-	ExtNum  string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int    `json:"accrual"`
+	ExtNum  string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float64 `json:"accrual"`
 }
